// Capstone Project/src/app/pages/landlord/host-house-detail-page/screen/types.ts
export enum hostHouseDetailPageEnum {
  GET_HOUSE_DETAIL_REQUEST = 'APP/HOST_HOUSE_LIST/GET_HOUSE_DETAIL_REQUEST',
  GET_HOUSE_DETAIL_SUCCESS = 'APP/HOST_HOUSE_LIST/GET_HOUSE_DETAIL_SUCCESS',
  GET_HOUSE_DETAIL_FAILURE = 'APP/HOST_HOUSE_LIST/GET_HOUSE_DETAIL_FAILURE',

  UPDATE_HOUSE_DETAIL_REQUEST = 'APP/HOST_HOUSE_LIST/UPDATE_HOUSE_DETAIL_REQUEST',
  UPDATE_HOUSE_DETAIL_SUCCESS = 'APP/HOST_HOUSE_LIST/UPDATE_HOUSE_DETAIL_SUCCESS',
  UPDATE_HOUSE_DETAIL_FAILURE = 'APP/HOST_HOUSE_LIST/UPDATE_HOUSE_DETAIL_FAILURE',

  GET_HOUSE_HISTORY_REQUEST = 'APP/HOST_HOUSE_LIST/GET_HOUSE_HISTORY_REQUEST',
  GET_HOUSE_HISTORY_SUCCESS = 'APP/HOST_HOUSE_LIST/GET_HOUSE_HISTORY_SUCCESS',
  GET_HOUSE_HISTORY_FAILURE = 'APP/HOST_HOUSE_LIST/GET_HOUSE_HISTORY_FAILURE',

  CLEAR_HOUSE_DETAIL_STATE = 'APP/HOST_HOUSE_LIST/HOUSE_DETAIL_STATE',
}